+++ thesis_core/src/main/java/thesis/core/serialization/world/UAVEntityConfig.java
@@ -10,7 +10,7 @@ public class UAVEntityConfig
 {
    private int type;
    private WorldCoordinate location;
-   private float orientation;
+   private double orientation;
 
    public UAVEntityConfig()
    {
@@ -34,12 +34,12 @@ public class UAVEntityConfig
       return location;
    }
 
-   public float getOrientation()
+   public double getOrientation()
    {
       return orientation;
    }
 
-   public void setOrientation(float angle)
+   public void setOrientation(double angle)
    {
       orientation = angle;
    }
